Extract predicate factory to new file
