+++ predicate/is_not_none_predicate.py
@@ -1,8 +1,7 @@
 from dataclasses import dataclass
-from functools import partial
 from typing import Final, override
 
-from predicate.predicate import Predicate, and_p
+from predicate.predicate import Predicate, and_p, predicate_partial
 
 
 @dataclass
@@ -24,5 +23,5 @@ is_not_none_p: Final[IsNotNonePredicate] = IsNotNonePredicate()
 """Return True if value is not None, otherwise False."""
 
 
-none_is_false_p = partial(and_p, is_not_none_p)
+none_is_false_p = predicate_partial(and_p, is_not_none_p)
 """Return False if value is None, otherwise the result of the predicate."""
